fixed checkstyle mistakes
