+++ src/main/java/core/basesyntax/model/Cat.java
