+++ src/main/java/core/basesyntax/model/Dog.java
